# api/sarcoma.openapi.yaml
openapi: 3.0.3
info:
  title: Sarcom FastTrack API
  version: 1.0.0
  description: Go WebAPI for Sarcom FastTrack patient handoff reports.
  contact:
    email: support@sarcoma-fasttrack.local
  license:
    name: CC BY 4.0
    url: https://creativecommons.org/licenses/by/4.0/
servers:
  - description: Local API
    url: http://localhost:8000
  - description: WAC Gateway API route
    url: /sarcoma-fasttrack-api
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/patients:
    get:
      tags: [patients]
      summary: List patients.
      operationId: listPatients
      responses:
        "200":
          description: Patient list.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PatientRead"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [patients]
      summary: Create a patient.
      operationId: createPatient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatientCreate"
      responses:
        "200":
          description: Created patient.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientRead"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/patients/{patientId}:
    parameters:
      - $ref: "#/components/parameters/PatientId"
    get:
      tags: [patients]
      summary: Get patient detail.
      operationId: getPatient
      responses:
        "200":
          description: Patient detail.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientRead"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [patients]
      summary: Update patient detail.
      operationId: updatePatient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatientCreate"
      responses:
        "200":
          description: Updated patient.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientRead"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [patients]
      summary: Delete patient.
      operationId: deletePatient
      responses:
        "204":
          description: Patient deleted.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/patients/{patientId}/name:
    parameters:
      - $ref: "#/components/parameters/PatientId"
    get:
      tags: [patients]
      summary: Get display name for a patient.
      operationId: getPatientName
      responses:
        "200":
          description: Patient display name.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NameResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/organizations:
    get:
      tags: [organizations]
      summary: List organizations.
      operationId: listOrganizations
      responses:
        "200":
          description: Organization list.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrganizationRead"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [organizations]
      summary: Create an organization.
      operationId: createOrganization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrganizationCreate"
      responses:
        "200":
          description: Created organization.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganizationRead"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/organizations/{organizationId}:
    parameters:
      - $ref: "#/components/parameters/OrganizationId"
    get:
      tags: [organizations]
      summary: Get organization detail.
      operationId: getOrganization
      responses:
        "200":
          description: Organization detail.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganizationRead"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [organizations]
      summary: Update organization detail.
      operationId: updateOrganization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrganizationCreate"
      responses:
        "200":
          description: Updated organization.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganizationRead"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [organizations]
      summary: Delete organization.
      operationId: deleteOrganization
      responses:
        "204":
          description: Organization deleted.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/organizations/{organizationId}/name:
    parameters:
      - $ref: "#/components/parameters/OrganizationId"
    get:
      tags: [organizations]
      summary: Get display name for an organization.
      operationId: getOrganizationName
      responses:
        "200":
          description: Organization display name.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NameResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PatientId:
      in: path
      name: patientId
      required: true
      schema:
        type: integer
...
    OrganizationId:
      in: path
      name: organizationId
      required: true
      schema:
        type: integer
  responses:
    BadRequest:
      description: Invalid request payload.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid bearer token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource was not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    ErrorResponse:
      type: object
      additionalProperties: true
      properties:
        error:
          type: string
        detail:
          type: string
...
    PatientCreate:
      type: object
      properties:
        first_name:
          type: string
        last_name:
          type: string
        given_name:
          type: string
        family_name:
          type: string
        address:
          type: string
        address_text:
          type: string
        birth_number:
          type: string
        identifier_rc:
          type: string
        phone:
          type: string
        email:
          type: string
          format: email
    PatientRead:
      allOf:
        - $ref: "#/components/schemas/PatientCreate"
        - type: object
          required: [id]
          properties:
            id:
              type: integer
            fhir_id:
              type: string
    OrganizationCreate:
      type: object
      properties:
        name:
          type: string
        type_code:
          type: string
        address:
          type: string
        contact:
          type: string
        email:
          type: string
          format: email
    OrganizationRead:
      allOf:
        - $ref: "#/components/schemas/OrganizationCreate"
        - type: object
          required: [id]
          properties:
            id:
              type: integer
            fhir_id:
              type: string
...
    NameResponse:
      type: object
      required: [name]
      properties:
        name:
          type: string